corrective actions

+++ app/(mgmt)/[lang]/deviations/actions.ts
@@ -4,15 +4,14 @@ import { auth } from '@/auth';
 import { dbc } from '@/lib/mongo';
 import { ObjectId } from 'mongodb';
 import { revalidateTag } from 'next/cache';
-import { redirect } from 'next/navigation';
+// import { redirect } from 'next/navigation';
 
 export async function deleteDraftDeviation(_id: ObjectId) {
+  const session = await auth();
+  if (!session || !session.user.email) {
+    return { error: 'unauthorized' };
+  }
   try {
-    const session = await auth();
-    if (!session) {
-      redirect('/auth');
-    }
-
     const collection = await dbc('deviations');
 
     const deviation = await collection.findOne({ _id: new ObjectId(_id) });
@@ -23,6 +22,10 @@ export async function deleteDraftDeviation(_id: ObjectId) {
 
     if (deviation.status !== 'draft') {
       return { error: 'not draft' };
+    }
+
+    if (deviation.owner !== session.user.email) {
+      return { error: 'unauthorized' };
     }
 
     const res = await collection.deleteOne({ _id: new ObjectId(_id) });
